added AppRegistry::uninstall and corresponding service API
also moved each app-service unit test into it's own top level crate, to
force `cargo test` to only run one test at a time, serially (they are
not currently thread safe)
